fix benchmark tests
