+++ Monorepo/ExampleAppEventSourcing/Symfony/config/services.php
@@ -14,7 +14,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $containerConfigurator->extension('ecotone', [
         'defaultSerializationMediaType' => 'application/json',
         'loadSrcNamespaces' => false,
-        'namespaces' => ['Monorepo\\ExampleAppEventSourcing\\Common\\'],
+        'namespaces' => ['Monorepo\\ExampleAppEventSourcing\\Common\\', 'Monorepo\\ExampleAppEventSourcing\\ProophProjection\\'],
         'defaultErrorChannel' => 'errorChannel',
         'failFast' => false,
         'skippedModulePackageNames' => \json_decode(\getenv('APP_SKIPPED_PACKAGES'), true),
@@ -23,6 +23,10 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
 
     $services->load('Monorepo\\ExampleAppEventSourcing\\Common\\', '%kernel.project_dir%/../Common/')
+        ->autowire()
+        ->autoconfigure();
+
+    $services->load('Monorepo\\ExampleAppEventSourcing\\ProophProjection\\', '%kernel.project_dir%/../ProophProjection/')
         ->autowire()
         ->autoconfigure();
